Moved most HDF5 functionality into it's own namespace
Moved low level HDF5 functionality to a seperate HDF5Utils file in the
utils directory. This has two main consequences. A: there are less
HDF5 C API calls within the FispactProblem.C source file itself, making
it slightly neater. B: Potentially easier testability of the HDF5
functionality, as different HDF5 mechanisms are now split into their own
functions.

As a result of reorganising the HDF5 reads/writes, one of the input
parameters has changed. Previously, to read the neutron flux spectra
from the OpenMC statepoint file, the user was asked to pass the internal
HDF5 path to the required dataset as an input parameter to
FispactProblem. This was janky. This has been changed so that now the
user only need pass in the OpenMC tally id representing the energy binnedneutron flux spectra.